packagekit: Split out the RestartServices modal
